requirements.txt
# Make sure clams-python version is explicitly specified, at least the lower bound
clams-python==1.2.6
pyannote.audio >=3.1.1
ffmpeg-python>=0.2.0
